// infoenergo-ui/src/components/ui/AiTabs/types.ts
import type { useTabsBaseOperations } from '@/composables/AiTabs/useTabsBaseOperations';
export interface Tab {
    /**
     * Отображаемое название таба в заголовке вкладки.
     */
    title: string;

    /**
     * Уникальный идентификатор представления (view), связанного с табом.
     * Используется для поиска, восстановления и сравнения вкладок.
     */
    viewId: string;

    /**
     * Имя роута, к которому привязан таб (из конфигурации роутера).
     */
    routeName: string;

    /**
     * Полный путь роута, который открывается при активации таба.
     * Хранится для возможности восстановления навигации.
     */
    routePath?: string;

    /**
     * Флаг, обозначающий, что данные, связанные с табом, были изменены
     * и отличаются от исходного/сохранённого состояния.
     */
    isModified: boolean;

    /**
     * Пользовательские данные, связанные с содержимым таба
     * (например, форма, выбранные фильтры и т.п.).
     */
    userData: any;

    /**
     * Флаг, определяющий, можно ли закрыть таб (показывать кнопку "Закрыть").
     */
    isClosable: boolean;

    /**
     * Параметры роута или дополнительные данные,
     * по которым можно однозначно идентифицировать вкладку.
     */
    params?: any;

    /**
     * Дополнительное имя расширения/сущности,
     * которое может переопределять стандартный заголовок таба.
     */
    extensionName?: string | null;
}

export interface TabState {
    /**
     * Активный таб
     */
    activeTab: Tab;
    /**
     * Массив табов
     */
    tabs: Tab[];
}

export type UseTabsBaseOperationsReturn = ReturnType<typeof useTabsBaseOperations>;

export interface AiTabsEmits {
    /**
     * Событие закрытия вкладки. Срабатывает, когда пользователь нажимает на кнопку закрытия вкладки. Возвращает название вкладки(поле *title*):
     */
    closeTab: [tabName: string];
    /**
     * Событие фокуса на вкладке. Срабатывает, когда происходит нажатие на вкладку и она становиться активной. Возвращает название вкладки(поле *title*):
     */
    focusTab: [tabName: string];
    /**
     * Событие клика по кнопке "Да" в диалоге.
     */
    onDialogPositiveClick: [tabName: string];
    /**
     * Событие клика по кнопке "Нет" в диалоге.
     */
    onDialogNegativeClick: [tabName: string];
    /**
     * Событие клика по кнопке "Отмена" в диалоге.
     */
    onDialogNeutralClick: [tabName: string];
    /**
     *
     * Событие вызвается после нажатия на пункт контекстного меню "Закрыть все вкладки".
     */
    closeAllTabs: [void];
    /**
     * Событие вызвается после нажатия на пункт контекстного меню "Закрыть все вкладки, кроме текущей".
     * Возвращает название вкладки, которую не нужно закрывать(поле *tabName*).
     */
    closeAllExceptCurrent: [string];
}